requirements.txt
fastapi==0.105.1
uvicorn==0.23.2
SQLAlchemy==2.0.20
psycopg2-binary==2.9.9
python-dotenv==1.0.0

